Add Express server at port 3001 to render components
The express server will take the requested path and serialized state
as input and will return the HTML representation as well as the state
embedded in script tags. You will want to use something like supervisor
to restart the server on component changes. Technically we could have
used Webpack to hotload the changes, but I couldn't figure it out