Server-side Render Components
We can now take advantage of the express server and "server-side"
render the components. The result is complete HTML pages being returned.
In fact, you should now be able to navigate the page without the
need for Javascript enabled, time-to-render on the client is faster,
obvious SEO benefits. The downside is a "round-trip" to the express
server